backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
requests
beautifulsoup4
python-dotenv
langchain
langchain-core
langchain-google-genai
gunicorn